sendrecv: fix MinGW-w64 warning
The first argument to select is an int, while curl_socket_t is
unsigned long long when using WinSock. It's ignored anyway [1].

[1] https://msdn.microsoft.com/en-us/library/windows/desktop/ms740141.aspx

+++ docs/examples/sendrecv.c
@@ -52,7 +52,7 @@ static int wait_on_socket(curl_socket_t sockfd, int for_recv, long timeout_ms)
   }
 
   /* select() returns the number of signalled sockets or -1 */
-  res = select(sockfd + 1, &infd, &outfd, &errfd, &tv);
+  res = select((int)sockfd + 1, &infd, &outfd, &errfd, &tv);
   return res;
 }
